agents/openai/1.0.0/requirements.txt
bs4==0.0.2
langchain-community==0.0.19
langchain_experimental==0.0.50
langchainhub==0.1.14
langchain_openai==0.0.5